updated donation voting merkle functions
