--- src/strategies/DonationVotingMerkleDistributionStrategy/donationVoting.config.ts
@@ -1,1225 +1,0 @@
-// uses Vault abi, since it's the same like direct transfer + an additional claim function
-export const abi = [
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "_allo",
-        type: "address",
-      },
-      {
-        internalType: "string",
-        name: "_name",
-        type: "string",
-      },
-      {
-        internalType: "contract ISignatureTransfer",
-        name: "_permit2",
-        type: "address",
-      },
-    ],
-    stateMutability: "nonpayable",
-    type: "constructor",
-  },
-  {
-    inputs: [],
-    name: "ALLOCATION_ACTIVE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ALLOCATION_NOT_ACTIVE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ALLOCATION_NOT_ENDED",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ALREADY_INITIALIZED",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "AMOUNT_MISMATCH",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ANCHOR_ERROR",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ARRAY_MISMATCH",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "INVALID",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "INVALID_ADDRESS",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "INVALID_FEE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "INVALID_METADATA",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "INVALID_REGISTRATION",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "IS_APPROVED_STRATEGY",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "MISMATCH",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NONCE_NOT_AVAILABLE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NOT_APPROVED_STRATEGY",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NOT_ENOUGH_FUNDS",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NOT_INITIALIZED",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "NOT_PENDING_OWNER",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "POOL_ACTIVE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "POOL_INACTIVE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "RECIPIENT_ALREADY_ACCEPTED",
-    type: "error",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-    ],
-    name: "RECIPIENT_ERROR",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "RECIPIENT_NOT_ACCEPTED",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "REGISTRATION_NOT_ACTIVE",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "UNAUTHORIZED",
-    type: "error",
-  },
-  {
-    inputs: [],
-    name: "ZERO_ADDRESS",
-    type: "error",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "token",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "Allocated",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "BatchPayoutSuccessful",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "recipientAddress",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "token",
-        type: "address",
-      },
-    ],
-    name: "Claimed",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "recipientAddress",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "Distributed",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "bytes32",
-        name: "merkleRoot",
-        type: "bytes32",
-      },
-      {
-        components: [
-          {
-            internalType: "uint256",
-            name: "protocol",
-            type: "uint256",
-          },
-          {
-            internalType: "string",
-            name: "pointer",
-            type: "string",
-          },
-        ],
-        indexed: false,
-        internalType: "struct Metadata",
-        name: "metadata",
-        type: "tuple",
-      },
-    ],
-    name: "DistributionUpdated",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "amount",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "grantee",
-        type: "address",
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "token",
-        type: "address",
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-    ],
-    name: "FundsDistributed",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "uint8",
-        name: "version",
-        type: "uint8",
-      },
-    ],
-    name: "Initialized",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "address",
-        name: "allo",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "bytes32",
-        name: "profileId",
-        type: "bytes32",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "poolId",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "bytes",
-        name: "data",
-        type: "bytes",
-      },
-    ],
-    name: "Initialized",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "bool",
-        name: "active",
-        type: "bool",
-      },
-    ],
-    name: "PoolActive",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "uint256",
-        name: "rowIndex",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "uint256",
-        name: "fullRow",
-        type: "uint256",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "RecipientStatusUpdated",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "bytes",
-        name: "data",
-        type: "bytes",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "Registered",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: "uint64",
-        name: "registrationStartTime",
-        type: "uint64",
-      },
-      {
-        indexed: false,
-        internalType: "uint64",
-        name: "registrationEndTime",
-        type: "uint64",
-      },
-      {
-        indexed: false,
-        internalType: "uint64",
-        name: "allocationStartTime",
-        type: "uint64",
-      },
-      {
-        indexed: false,
-        internalType: "uint64",
-        name: "allocationEndTime",
-        type: "uint64",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-    ],
-    name: "TimestampsUpdated",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "bytes",
-        name: "data",
-        type: "bytes",
-      },
-      {
-        indexed: false,
-        internalType: "address",
-        name: "sender",
-        type: "address",
-      },
-      {
-        indexed: false,
-        internalType: "uint8",
-        name: "status",
-        type: "uint8",
-      },
-    ],
-    name: "UpdatedRegistration",
-    type: "event",
-  },
-  {
-    inputs: [],
-    name: "NATIVE",
-    outputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "PERMIT2",
-    outputs: [
-      {
-        internalType: "contract ISignatureTransfer",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "bytes",
-        name: "_data",
-        type: "bytes",
-      },
-      {
-        internalType: "address",
-        name: "_sender",
-        type: "address",
-      },
-    ],
-    name: "allocate",
-    outputs: [],
-    stateMutability: "payable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "allocationEndTime",
-    outputs: [
-      {
-        internalType: "uint64",
-        name: "",
-        type: "uint64",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "allocationStartTime",
-    outputs: [
-      {
-        internalType: "uint64",
-        name: "",
-        type: "uint64",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "allowedTokens",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        components: [
-          {
-            internalType: "address",
-            name: "recipientId",
-            type: "address",
-          },
-          {
-            internalType: "address",
-            name: "token",
-            type: "address",
-          },
-        ],
-        internalType:
-          "struct DonationVotingMerkleDistributionVaultStrategy.Claim[]",
-        name: "_claims",
-        type: "tuple[]",
-      },
-    ],
-    name: "claim",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "claims",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address[]",
-        name: "_recipientIds",
-        type: "address[]",
-      },
-      {
-        internalType: "bytes",
-        name: "_data",
-        type: "bytes",
-      },
-      {
-        internalType: "address",
-        name: "_sender",
-        type: "address",
-      },
-    ],
-    name: "distribute",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "distributionMetadata",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "protocol",
-        type: "uint256",
-      },
-      {
-        internalType: "string",
-        name: "pointer",
-        type: "string",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "distributionStarted",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "getAllo",
-    outputs: [
-      {
-        internalType: "contract IAllo",
-        name: "",
-        type: "address",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address[]",
-        name: "_recipientIds",
-        type: "address[]",
-      },
-      {
-        internalType: "bytes[]",
-        name: "_data",
-        type: "bytes[]",
-      },
-    ],
-    name: "getPayouts",
-    outputs: [
-      {
-        components: [
-          {
-            internalType: "address",
-            name: "recipientAddress",
-            type: "address",
-          },
-          {
-            internalType: "uint256",
-            name: "amount",
-            type: "uint256",
-          },
-        ],
-        internalType: "struct IStrategy.PayoutSummary[]",
-        name: "",
-        type: "tuple[]",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "getPoolAmount",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "getPoolId",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "_recipientId",
-        type: "address",
-      },
-    ],
-    name: "getRecipient",
-    outputs: [
-      {
-        components: [
-          {
-            internalType: "bool",
-            name: "useRegistryAnchor",
-            type: "bool",
-          },
-          {
-            internalType: "address",
-            name: "recipientAddress",
-            type: "address",
-          },
-          {
-            components: [
-              {
-                internalType: "uint256",
-                name: "protocol",
-                type: "uint256",
-              },
-              {
-                internalType: "string",
-                name: "pointer",
-                type: "string",
-              },
-            ],
-            internalType: "struct Metadata",
-            name: "metadata",
-            type: "tuple",
-          },
-        ],
-        internalType:
-          "struct DonationVotingMerkleDistributionBaseStrategy.Recipient",
-        name: "recipient",
-        type: "tuple",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "_recipientId",
-        type: "address",
-      },
-    ],
-    name: "getRecipientStatus",
-    outputs: [
-      {
-        internalType: "enum IStrategy.Status",
-        name: "",
-        type: "uint8",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "getStrategyId",
-    outputs: [
-      {
-        internalType: "bytes32",
-        name: "",
-        type: "bytes32",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_index",
-        type: "uint256",
-      },
-    ],
-    name: "hasBeenDistributed",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_amount",
-        type: "uint256",
-      },
-    ],
-    name: "increasePoolAmount",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_poolId",
-        type: "uint256",
-      },
-      {
-        internalType: "bytes",
-        name: "_data",
-        type: "bytes",
-      },
-    ],
-    name: "initialize",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "isDistributionSet",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "isPoolActive",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "_allocator",
-        type: "address",
-      },
-    ],
-    name: "isValidAllocator",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "merkleRoot",
-    outputs: [
-      {
-        internalType: "bytes32",
-        name: "",
-        type: "bytes32",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "metadataRequired",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "bytes[]",
-        name: "data",
-        type: "bytes[]",
-      },
-    ],
-    name: "multicall",
-    outputs: [
-      {
-        internalType: "bytes[]",
-        name: "results",
-        type: "bytes[]",
-      },
-    ],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "address",
-        name: "",
-        type: "address",
-      },
-    ],
-    name: "recipientToStatusIndexes",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "recipientsCounter",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "bytes",
-        name: "_data",
-        type: "bytes",
-      },
-      {
-        internalType: "address",
-        name: "_sender",
-        type: "address",
-      },
-    ],
-    name: "registerRecipient",
-    outputs: [
-      {
-        internalType: "address",
-        name: "recipientId",
-        type: "address",
-      },
-    ],
-    stateMutability: "payable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "registrationEndTime",
-    outputs: [
-      {
-        internalType: "uint64",
-        name: "",
-        type: "uint64",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "registrationStartTime",
-    outputs: [
-      {
-        internalType: "uint64",
-        name: "",
-        type: "uint64",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        components: [
-          {
-            internalType: "uint256",
-            name: "index",
-            type: "uint256",
-          },
-          {
-            internalType: "uint256",
-            name: "statusRow",
-            type: "uint256",
-          },
-        ],
-        internalType:
-          "struct DonationVotingMerkleDistributionBaseStrategy.ApplicationStatus[]",
-        name: "statuses",
-        type: "tuple[]",
-      },
-    ],
-    name: "reviewRecipients",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    name: "statusesBitMap",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "totalPayoutAmount",
-    outputs: [
-      {
-        internalType: "uint256",
-        name: "",
-        type: "uint256",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "bytes32",
-        name: "_merkleRoot",
-        type: "bytes32",
-      },
-      {
-        components: [
-          {
-            internalType: "uint256",
-            name: "protocol",
-            type: "uint256",
-          },
-          {
-            internalType: "string",
-            name: "pointer",
-            type: "string",
-          },
-        ],
-        internalType: "struct Metadata",
-        name: "_distributionMetadata",
-        type: "tuple",
-      },
-    ],
-    name: "updateDistribution",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint64",
-        name: "_registrationStartTime",
-        type: "uint64",
-      },
-      {
-        internalType: "uint64",
-        name: "_registrationEndTime",
-        type: "uint64",
-      },
-      {
-        internalType: "uint64",
-        name: "_allocationStartTime",
-        type: "uint64",
-      },
-      {
-        internalType: "uint64",
-        name: "_allocationEndTime",
-        type: "uint64",
-      },
-    ],
-    name: "updatePoolTimestamps",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    inputs: [],
-    name: "useRegistryAnchor",
-    outputs: [
-      {
-        internalType: "bool",
-        name: "",
-        type: "bool",
-      },
-    ],
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    inputs: [
-      {
-        internalType: "uint256",
-        name: "_amount",
-        type: "uint256",
-      },
-    ],
-    name: "withdraw",
-    outputs: [],
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    stateMutability: "payable",
-    type: "receive",
-  },
-];
