+++ src/strategies/DonationVotingMerkleDistributionStrategy/types.ts
@@ -11,6 +11,12 @@ export type Recipient = {
   metadata: Metadata;
 };
 
+export type RegisterDataDonationVoting = {
+  registryAnchor: `0x${string}`;
+  recipientAddress: `0x${string}`;
+  metadata: Metadata;
+};
+
 export type Distribution = {
   index: number;
   recipientId: string;
@@ -18,17 +24,33 @@ export type Distribution = {
   merkleProof: string[];
 };
 
-export type InitializeData = {
+export type InitializeParamsDonationVoting = {
   useRegistryAnchor: boolean;
   metadataRequired: boolean;
-  registrationStartTime: number;
-  registrationEndTime: number;
-  allocationStartTime: number;
-  allocationEndTime: number;
-  allowedTokens: string[];
+  registrationStartTime: bigint;
+  registrationEndTime: bigint;
+  allocationStartTime: bigint;
+  allocationEndTime: bigint;
+  allowedTokens: `0x${string}`[];
+};
+
+export type TokenPermissions = {
+  token: `0x${string}`;
+  amount: bigint;
+};
+
+export type PermitTransferFrom = {
+  permitted: TokenPermissions;
+  nonce: bigint;
+  deadline: bigint;
 };
 
 export type Permit2Data = {
-  permit: any;
-  signature: string;
+  permit: PermitTransferFrom;
+  signature: `0x${string}`;
+};
+
+export const StrategyTypeDonationVoting = {
+  Vault: "DonationVotingMerkleDistributionVaultv1",
+  Direct: "DonationVotingMerkleDistributionDirectTransferv1",
 };
